исправил передачу данных
