+++ js/data/userName.js
@@ -1,6 +1,6 @@
 export default (state, userName) => {
-  if (!userName) {
-    throw new RangeError(`Username can not be empty`);
+  if (typeof userName !== `string` && userName === ``) {
+    throw new RangeError(`Username can not be empty or not a string type`);
   }
   return Object.assign({}, state, {'userName': userName});
 };
