+++ js/game/game-model.js
@@ -1,10 +1,6 @@
 import Model from '../model';
 
 export default new class extends Model {
-  constructor() {
-    super();
-    this.games = [];
-  }
   get urlRead() {
     return `https://intensive-ecmascript-server-btfgudlkpi.now.sh/pixel-hunter/questions`;
   }
