+++ js/intro/intro.js
@@ -1,6 +1,5 @@
 import switchDisplay from '../switch-display';
 import IntroView from './intro-view';
-import App from '../main';
 
 export default class Intro {
   constructor() {
@@ -9,9 +8,5 @@ export default class Intro {
 
   init() {
     switchDisplay(this.view);
-
-    this.view.onChangeDisplay = () => {
-      App.showGreeting();
-    };
   }
 }
